+++ lib/src/command/lib.dart
@@ -47,21 +47,8 @@ class LibCommand extends BaseVoidCommand {
     }
     await lib.download();
 
-    final compiler = _createCompiler(lib);
+    final compiler = createCompiler(lib);
     await compiler.compile(lib);
   }
 
-  BaseCompiler _createCompiler(Lib lib) {
-    final type = lib.type;
-    switch (type) {
-      case LibType.cAutotools:
-        return AutoToolsCompiler();
-      case LibType.cCmake:
-        return CMakeCompiler();
-      case LibType.cMeson:
-        return MesonCompiler();
-      case LibType.cMakefile:
-        return MakefileCompiler();
-    }
-  }
 }