+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UnRelease
 
+- Add `rust` compiler.
+- Refactored `flags`.
+- Fix meson build error with flag.
+
 ## 0.0.6
 
 - Rename `compile` command to `lib`.